Fix issue with rolling update when some existing replicas are unhealthy
The bug is as follows:
* assume replicas readiness is [NOT_READY, READY]
* update is triggered, with MaxUnavailable=1
* update starts updating the replica-1, causing complete downtime,
  while the desired behavior would be to wait for replica-0 to become ready

The fix is to reduce partition less aggressively - taking into account pods
that are not ready.

Note that the fix was tested on a cluster with MaxUnavailableStatefulSet feature disabled
(I don't have access to a cluster with this feature enabled, maybe the issue is less impactful there,
however arguably still beneficial since the lws considers readiness of leader+workers and not only leader).